backend/requirements.txt
django==4.0.1
psycopg2==2.9.3
djangorestframework==3.13.1
django-cors-headers==3.11.0
drf-spectacular==0.22.0
djangorestframework-simplejwt==5.1.0
django-guardian==2.4.0
django-filter==21.1
django-crispy-forms==1.14.0